Clean up shebang line and add test for it.
Needed shebang line test as in the first pass at adding the
upload-cghub-bam script this was not correct and it was not useable as
a stand-alone program. Note: the Test::Script::Shebang is NOT
acceptable (fails to recognize #! /usr/bin/env perl as legitimate, and
(from reviewing source code) appears to allow spaces before it on the
first line.