generate data test passes.
Modified code a bit, changed the analysis alias to something useful.
This is a CONTENT change, but the old value did not allow looking it up
in our system or in any way referencing it after release, other than by
searching through all files in all directories for a particular value.